refactor: relocate claim limiter and tidy docs
